Fix vertical mode tests subclassing main test
